Init project
We add test scripts for the noninteracting case, in both 1D and 2D.
These results are for zero temperature and, in 1D, match published
findings [1], in the thermodynamic limit.

We aim to generalize the script to finite temperature to witness a
thermal death of entanglement, as well as the spatial one.

[1] L. Ding, G. Dünnweber, C. Schilling, 10.1088/2058-9565/ad00d9,
    Quantum Science and Technology 9 (1), 015005