Swap to non-logarithmic negativities for the 1D test
We lose the LNE ≥ QRE condition, as N is sometimes larger and sometimes
smaller than the QRE (at least under sym-resolution / superselection...).
Specifically, we find N > E close to half filling and N < E at smaller
fillings, with N = E at some intermediate filling, which is dependent on
the level of SSR (charge or parity) and also on the distance between the
two sites. So, much less general and terse. But, overall, the agreement
between the two quantities is better, as they "live on the same scale",
for the whole analysis and look very similar if we plot two panels side
by side. I might go back on this, but for now let's swap.