Strengthen the conditions for the analytical formulas for the SSR-QRE
They are really meant for exact symmetries, ouch. This accounts for the
small differences between the analytical and the semi-numerical formula,
in our CDMFT data for the Mott transition. Unfortunately at the time we
ran the calculation for the Phys. Rev. B 109, 115104 (2024) paper, we
implemented the analytical formulas, as found in [1], instead of the
semi-numerical formulation originally derived in the appendix of [2].

[1] L. Ding, G. Dünnweber, C. Schilling, 10.1088/2058-9565/ad00d9,
    Quantum Science and Technology 9 (1), 015005

[2] Lexin Ding, Zoltan Zimboras, Christian Schilling, arXiv:2207.03377